+++ spec/models/bigbluebutton_recording_db_spec.rb
@@ -16,7 +16,7 @@ describe BigbluebuttonRecording do
     it { should have_db_column(:description).of_type(:string) }
     it { should have_db_column(:created_at).of_type(:datetime) }
     it { should have_db_column(:updated_at).of_type(:datetime) }
-    it { should have_db_column(:size).of_type(:integer) }
+    it { should have_db_column(:size).of_type(:bigint) }
     it { should have_db_index(:room_id) }
     it { should have_db_index(:recordid).unique(true) }
     it "default values" do
